Revise Docker Strategy
Instead of having a large monolithic docker file I have elected to
use docker-compose to support the database.  I've also decided that
environment variables should drive the configuration over having
them defined in the files.  This allows you to set them with docker
later when running it in a production like environment.