Add Simple Sign-Up and Sign-In
This is an introductory set of changes to open the door to letting
a user sign-up and sign-in.  I've drew a lot of this from this:

http://meatherly.github.io/2015/05/11/phoenixauthentication/

This just barely gets the ball rolling; however, it is a good
working start.  Currently this throws the user into a session but
I think the next best step is to add `guardian` to the project:

https://github.com/ueberauth/guardian