Add Guardian to Session Stack
This introduces Guardian as the authentication token.  It keeps a
JSON Web Token (JWT) browser side and stores the user details
encrypted in it.  The browser pipeline looks for this token, get's
the information from the token and pulls the user with it.

I've decided to use Guardian as it appears to be a good system
to store users auth token and is well documented.

I've also slipped in a `signed_in?` detection and have added a
"sign-in/sign-out" button that's based on your logged in status.